refactor(behaviors): Clean up caps_word
Renamed macros that still referred to "break" instead of "continuation".

Changed the type of the continuations count field to allow for more than
255 continuation keys.

Reordered a struct to allow for better packing.

Switched to using one static variable instead of a field in every
caps_word device to initialize the devs[] array.

Refactored caps_word_enhance_usage() to allow for more easily changing
the conditions for whether mods should be applied in future commits.